split up examples
